moved custom types into separate file
